some more fixup regarding integration with sly
I was trying to use breeze.listener:rcp-eval to load breeze for the
first time -_- d'oh

This commits add a lower-level function breeze-%eval that takes a
form (a list) instead of a string. This is now the function that take
care of choosing between sly or slime.

This caused issues with the readers (emacs lisp and cl readers) for
the "complex" form used to load breeze's system. So I extracted that
form into the src/ensure-breeze.lisp file.