WIP trying to use patterns to match more complex parse trees
P.S. This breaks a lot of tests because the "bindings" are still a
plist, but the keys are the names of the terms instead of the "term
pattern object" itself.

Returning the name instead of the pattern object is _much_ more
convenient (so far), but it means that we must be very very careful if
we want to re-use and nest patterns.